All linter, tests and stylers happy
